requirements.txt
selenium
PyGithub
python-gitlab

